removed everything except sd_vector_builder_disk
